Bug fix: remove template, settings, async download template

+++ src/utils/projectsManager.js
@@ -1,7 +1,8 @@
-import { notify } from "power-notifier";
+import {notify} from "power-notifier";
 import TemplatesRepo from "../stores/templatesRepo";
+import ProjectsRepo from "../stores/projectsRepo";
 import {generateIdFromName} from "./utils";
-import {childProcess, fse, _path} from "../requires";
+import {_path, fse} from "../requires";
 
 const TEMPLATES_PATH = "/.Pemplates/templates/";
 const PROJECTS_PATH = "/.Pemplates/projects/";
@@ -10,26 +11,33 @@ export const removeProject = ({ path }) => {
 	fse.removeSync(path);
 };
 
-export const createProjectFromTemplate = async ({ id }) => {
+export const createProjectFromTemplate = async ({ name, id }) => {
+	if(ProjectsRepo.has(name)){
+		if(!confirm("Project with this name already exists, would you like to override it?"))
+			return false;
+	}
+	
 	const template = TemplatesRepo.getTemplate(id);
 	//todo: projects with same names will be overridden
-	const targetPath = _path.join(__dirname, PROJECTS_PATH, template.name);
+	const targetPath = _path.join(__dirname, PROJECTS_PATH, name);
 	const srcPath = _path.join(__dirname, TEMPLATES_PATH, template.name);
 	
 	await fse.ensureDirSync(targetPath);
 	await fse.copySync(srcPath, targetPath);
 	
-	notify({
-		title: "New project created",
-		timeout: 2500
-	});
-	
 	const project = {
 		path: targetPath,
-		name: template.name,
+		name: name,
 		id: generateIdFromName(template.name),
 		templateId: id
 	};
 	
+	await ProjectsRepo.addProject(project);
+	
+	notify({
+		title: "New project created",
+		timeout: 2500
+	});
+
 	return project;
 };
